feat: fix template modal
